Character Table Lists Characters
Added a table which lists each PC. The View, Edit, and Delete buttons do not work.
Changed the colors. The old colors are commented out and can easily be returned.
The Feats chart has a scroll bar and does not overflow anymore.
Applied the white bg to the character and spells menu.